Module dropbox_backup_to_external_disk::local_mod

Module contains all functions for local external disk.

Enums

RemoteKind

The source file can be on dropbox or on external disk Backup_1

Functions

add2_just_downloaded_to_list_local

add just downloaded files to list_local (from external disk)

add_just_downloaded_to_list_local

add just downloaded files to list_local (from dropbox remote)

add_just_downloaded_to_list_local_internal

add lines from just_downloaded to list_local. Only before compare.

copy_from_list2_for_download

copies files from external disk backup_1 to backup_2

correct_time_from_list

modify the date od files from list_for_correct_time

correct_time_from_list_internal

modify the date od files from list_for_correct_time

list2_local

for second backup: list all local files and folders. It can take some time.

list_local

list all local files and folders. It can take some time.

list_local_internal

list all local files and folders. It can take some time.

move_internal

internal code to move file

move_or_rename_local_files

Files are often moved or renamed
After compare, the same file (with different path or name) will be in the list_for_trash and in the list_for_download.
First for every trash line, we search list_for_download for same size and modified.
If found, get the remote_metadata with content_hash and calculate local_content_hash.
If they are equal move or rename, else nothing: it will be trashed and downloaded eventually.
Remove also the lines in files list_for_trash and list_for_download.

move_or_rename_local_files_internal

internal function

save2_base_path

saves the base local path for later use like “/mnt/f/DropBoxBackup2”

save_base_path

saves the base local path for later use like “/mnt/d/DropBoxBackup1”

trash2_from_list

Move to trash folder the files from list_for_trash.
Ignore if the file does not exist anymore.

trash_from_list

Move to trash folder the files from list_for_trash.
Ignore if the file does not exist anymore.

trash_from_list_internal

internal