Module dropbox_backup_to_external_disk::local_mod

local_mod.rs Module contains all functions for local external disk.

Functions

add_just_downloaded_to_list_local
add lines from just_downloaded to list_local

correct_time_from_list
modify the files from list_for_correct_time

list_local
list all local files and folders. It can take some time.

move_or_rename_local_files
Files are often moved or renamed After compare, the same file (with different path or name) will be in the list_for_trash and in the list_for_download. First for every trash line, we search list_for_download for same size and modified. If found, get the remote_metadata with content_hash and calculate local_content_hash. If they are equal move or rename, else leave: it will be trashed and downloaded eventually. Remove also the lines in files list_for_trash and list_for_download.

save_base_path
save the base local path for later commands

trash_from_list
move to trash folder the files from list_for_trash