Module dropbox_backup_to_external_disk::remote_mod

remote_mod.rs Module contains all the communication with the remote dropbox storage.

Structs

DirectoryIterator
iterator for Directory on remote Dropbox storage

Functions

download
download one file

download_from_list
download files from list

download_with_client
download one file with client object UserAuthDefaultClient

get_short_lived_access_token
get token from env variable

list_directory
list directory

list_remote
get remote list in parallel first get the first level of folders and then request in parallel sub-folders recursively

list_remote_folder
list remote folder

remote_content_hash
get content_hash from remote

sort_remote_list_and_write_to_file
sort and write to file

test_connection
test authentication with dropbox.com