Module dropbox_backup_to_external_disk::utils_mod

utils_mod.rs A module with often used functions. For every project I select only the functions I need for the project.

Functions

clear_line
ansi terminal - clears the line on the terminal from cursor to end of line

hide_cursor
ansi terminal - hide cursor

ns_elapsed
returns the elapsed nanoseconds

ns_print_ms
print elapsed time in milliseconds and returns the new now in nanoseconds

ns_print_ns
print elapsed time in nanoseconds and returns the new now in nanoseconds

ns_start
returns the now in nanoseconds

sort_string_lines

unhide_cursor
ansi terminal - unhide cursor