Module dropbox_backup_to_external_disk::utils_mod

utils_mod.rs A module with often used functions. For every project I select only the functions I need for the project.

Structs

CLEAR_ALL

CLEAR_LINE
ansi terminal - clears the line on the terminal from cursor to end of line

GREEN

HIDE_CURSOR

RESET

UNHIDE_CURSOR

YELLOW

Functions

at_line

at_pos

get_pos

ns_elapsed
returns the elapsed nanoseconds

ns_print_ms
print elapsed time in milliseconds and returns the new now in nanoseconds

ns_print_ns
print elapsed time in nanoseconds and returns the new now in nanoseconds

ns_start
returns the now in nanoseconds

sort_string_lines
sort string lines case insensitive

start_mouse_terminal